refactor: extract TrainingSession component from programBuilder screen
